koji support
- move everything related to areas out of `config.js` and into `areas.js` because of async behavior
- add `kojiOptions` to config
- if you set `map.general.geoJsonFileName` to the full koji url, it will now call it from there instead of reading from a json
