+++ server/src/configs/default.json
@@ -99,6 +99,9 @@
       "invasions": "https://raw.githubusercontent.com/WatWowMap/event-info/main/grunts/formatted.json",
       "masterfile": "https://raw.githubusercontent.com/WatWowMap/Masterfile-Generator/master/master-latest-react-map.json",
       "translations": "https://raw.githubusercontent.com/WatWowMap/pogo-translations/master"
+    },
+    "kojiOptions": {
+      "bearerToken": "KOJI_SECRET"
     }
   },
   "multiDomains": [],
