+++ server/src/services/areas.js
@@ -1,7 +1,80 @@
 /* eslint-disable no-console */
-const config = require('./config')
+const { default: center } = require('@turf/center')
+const fs = require('fs')
+const { resolve } = require('path')
 
-const loadAreas = () => {
+const config = require('./config')
+const fetchJson = require('./api/fetchJson')
+
+const manualGeojson = {
+  type: 'FeatureCollection',
+  features: config.manualAreas
+    .filter((area) =>
+      ['lat', 'lon', 'name'].every((k) => k in area && !area.hidden),
+    )
+    .map((area) => {
+      const { lat, lon, ...rest } = area
+      return {
+        type: 'Feature',
+        properties: {
+          center: [lat, lon],
+          manual: true,
+          key: rest.parent ? `${rest.parent}-${rest.name}` : rest.name,
+          ...rest,
+        },
+        geometry: {
+          type: 'Polygon',
+          coordinates: [[[lon, lat]]],
+        },
+      }
+    }),
+}
+
+const getGeojson = async (location) => {
+  if (location.startsWith('http')) {
+    console.log('Loading Kōji URL', location)
+    return fetchJson(
+      location,
+      {
+        headers: {
+          Authorization: `Bearer ${config.api.kojiOptions.bearerToken}`,
+        },
+      },
+      true,
+    ).then((res) => res?.data || { features: [] })
+  }
+  if (fs.existsSync(resolve(`${__dirname}/../configs/${location}`))) {
+    return JSON.parse(
+      fs.readFileSync(resolve(__dirname, `../configs/${location}`)),
+    )
+  }
+  return { features: [] }
+}
+
+// Load each areas.json
+const loadScanPolygons = async (fileName, domain) => {
+  const geojson = await getGeojson(fileName)
+  return {
+    ...geojson,
+    features: [
+      ...manualGeojson.features.filter(
+        (f) => !f.properties.domain || f.properties.domain === domain,
+      ),
+      ...geojson.features.map((f) => ({
+        ...f,
+        properties: {
+          ...f.properties,
+          key: f.properties.parent
+            ? `${f.properties.parent}-${f.properties.name}`
+            : f.properties.name,
+          center: center(f).geometry.coordinates.reverse(),
+        },
+      })),
+    ].sort((a, b) => a.properties.name.localeCompare(b.properties.name)),
+  }
+}
+
+const loadAreas = async () => {
   try {
     const normalized = { type: 'FeatureCollection', features: [] }
     Object.values(config.scanAreas).forEach((area) => {
@@ -21,7 +94,7 @@ const loadAreas = () => {
   }
 }
 
-const parseAreas = (areasObj) => {
+const parseAreas = async (areasObj) => {
   const polygons = {}
   const names = []
   const withoutParents = {}
@@ -54,8 +127,76 @@ const parseAreas = (areasObj) => {
   return { names, withoutParents, polygons }
 }
 
-const raw = loadAreas()
-const { names, withoutParents, polygons } = parseAreas(raw)
+// Check if an areas.json exists
+const { raw, names, withoutParents, polygons } = (async () => {
+  config.scanAreas = {
+    main: await loadScanPolygons(config.map.geoJsonFileName),
+    ...Object.fromEntries(
+      await Promise.all(
+        config.multiDomains.map(async (d) => [
+          d.general?.geoJsonFileName ? d.domain : 'main',
+          await loadScanPolygons(
+            d.general?.geoJsonFileName || config.map.geoJsonFileName,
+          ),
+        ]),
+      ),
+    ),
+  }
+
+  config.scanAreasMenu = Object.fromEntries(
+    Object.entries(config.scanAreas).map(([domain, areas]) => {
+      const parents = { '': { children: [], name: '' } }
+
+      const noHidden = {
+        ...areas,
+        features: areas.features.filter((f) => !f.properties.hidden),
+      }
+      // Finds unique parents and determines if the parents have their own properties
+      noHidden.features.forEach((feature) => {
+        if (feature.properties.parent) {
+          parents[feature.properties.parent] = {
+            name: feature.properties.parent,
+            details: areas.features.find(
+              (area) => area.properties.name === feature.properties.parent,
+            ),
+            children: [],
+          }
+        }
+      })
+
+      // Finds the children of each parent
+      noHidden.features.forEach((feature) => {
+        if (feature.properties.parent) {
+          parents[feature.properties.parent].children.push(feature)
+        } else if (!parents[feature.properties.name]) {
+          parents[''].children.push(feature)
+        }
+      })
+
+      // Create blanks for better formatting when there's an odd number of children
+      Object.values(parents).forEach(({ children }) => {
+        if (children.length % 2 === 1) {
+          children.push({
+            type: 'Feature',
+            properties: { name: '', manual: !!config.manualAreas.length },
+          })
+        }
+      })
+      return [
+        domain,
+        Object.values(parents).sort((a, b) => a.name.localeCompare(b.name)),
+      ]
+    }),
+  )
+  config.scanAreasObj = Object.fromEntries(
+    Object.values(config.scanAreas)
+      .flatMap((areas) => areas.features)
+      .map((feature) => [feature.properties.name, feature]),
+  )
+})().then(async () => ({
+  raw: await loadAreas(),
+  ...(await parseAreas(raw)),
+}))
 
 module.exports = {
   raw,
