+++ server/src/services/config.js
@@ -4,7 +4,6 @@ process.env.NODE_CONFIG_DIR = `${__dirname}/../configs`
 const fs = require('fs')
 const { resolve } = require('path')
 const dotenv = require('dotenv')
-const { default: center } = require('@turf/center')
 
 dotenv.config()
 
@@ -297,119 +296,6 @@ config.authMethods = [
     )
   }
 })
-
-const manualGeojson = {
-  type: 'FeatureCollection',
-  features: config.manualAreas
-    .filter((area) =>
-      ['lat', 'lon', 'name'].every((k) => k in area && !area.hidden),
-    )
-    .map((area) => {
-      const { lat, lon, ...rest } = area
-      return {
-        type: 'Feature',
-        properties: {
-          center: [lat, lon],
-          manual: true,
-          key: rest.parent ? `${rest.parent}-${rest.name}` : rest.name,
-          ...rest,
-        },
-        geometry: {
-          type: 'Polygon',
-          coordinates: [[[lon, lat]]],
-        },
-      }
-    }),
-}
-
-// Load each areas.json
-const loadScanPolygons = (fileName, domain) => {
-  const geojson = fs.existsSync(resolve(`${__dirname}/../configs/${fileName}`))
-    ? JSON.parse(fs.readFileSync(resolve(__dirname, `../configs/${fileName}`)))
-    : { features: [] }
-  return {
-    ...geojson,
-    features: [
-      ...manualGeojson.features.filter(
-        (f) => !f.properties.domain || f.properties.domain === domain,
-      ),
-      ...geojson.features.map((f) => ({
-        ...f,
-        properties: {
-          ...f.properties,
-          key: f.properties.parent
-            ? `${f.properties.parent}-${f.properties.name}`
-            : f.properties.name,
-          center: center(f).geometry.coordinates.reverse(),
-        },
-      })),
-    ].sort((a, b) => a.properties.name.localeCompare(b.properties.name)),
-  }
-}
-
-// Check if an areas.json exists
-config.scanAreas = {
-  main: loadScanPolygons(config.map.geoJsonFileName),
-  ...Object.fromEntries(
-    config.multiDomains.map((d) => [
-      d.general?.geoJsonFileName ? d.domain : 'main',
-      loadScanPolygons(
-        d.general?.geoJsonFileName || config.map.geoJsonFileName,
-      ),
-    ]),
-  ),
-}
-
-config.scanAreasMenu = Object.fromEntries(
-  Object.entries(config.scanAreas).map(([domain, areas]) => {
-    const parents = { '': { children: [], name: '' } }
-
-    const noHidden = {
-      ...areas,
-      features: areas.features.filter((f) => !f.properties.hidden),
-    }
-    // Finds unique parents and determines if the parents have their own properties
-    noHidden.features.forEach((feature) => {
-      if (feature.properties.parent) {
-        parents[feature.properties.parent] = {
-          name: feature.properties.parent,
-          details: areas.features.find(
-            (area) => area.properties.name === feature.properties.parent,
-          ),
-          children: [],
-        }
-      }
-    })
-
-    // Finds the children of each parent
-    noHidden.features.forEach((feature) => {
-      if (feature.properties.parent) {
-        parents[feature.properties.parent].children.push(feature)
-      } else if (!parents[feature.properties.name]) {
-        parents[''].children.push(feature)
-      }
-    })
-
-    // Create blanks for better formatting when there's an odd number of children
-    Object.values(parents).forEach(({ children }) => {
-      if (children.length % 2 === 1) {
-        children.push({
-          type: 'Feature',
-          properties: { name: '', manual: !!config.manualAreas.length },
-        })
-      }
-    })
-    return [
-      domain,
-      Object.values(parents).sort((a, b) => a.name.localeCompare(b.name)),
-    ]
-  }),
-)
-config.scanAreasObj = Object.fromEntries(
-  Object.values(config.scanAreas)
-    .flatMap((areas) => areas.features)
-    .map((feature) => [feature.properties.name, feature]),
-)
 
 config.api.pvp.leagueObj = Object.fromEntries(
   config.api.pvp.leagues.map((league) => [league.name, league.cp]),
